refactor(ruborist): move override matching out of graph.rs
The matching half of npm overrides (collect_parent_chain,
check_override, matches_parent_chain_for_rule) and its ten tests now
live in model/override_match.rs next to the parsing half in
override_rule.rs; graph.rs drops from 1187 to 743 lines and is back to
being the graph data structure. Pure move (same impl DependencyGraph
methods); the overrides fields become pub(crate) for the sibling
module.
