+++ crates/ruborist/src/model/mod.rs
@@ -122,6 +122,7 @@ pub mod git;
 pub mod graph;
 pub mod manifest;
 pub mod node;
+pub mod override_match;
 pub mod override_rule;
 pub mod package_json;
 pub mod package_lock;
